# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Projeto de PDWA5",
        "description": "API do projeto de PDWA5",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api",
            "description": "Endpoint da API."
        }
    ],
    "paths": {
...
        "/menu": {
            "get": {
                "tags": [
                    "Menu"
                ],
                "summary": "Listar menus.",
                "description": "Lista os menus.",
                "operationId": "414b868686e4a54d1aa46b4fc14acb61",
                "responses": {
                    "200": {
                        "description": "Menus lisitados com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": "1"
                                            },
                                            "user": {
                                                "type": "integer",
                                                "example": "1"
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Novo Menu"
                                            },
                                            "permalink": {
                                                "type": "string",
                                                "example": "novo-menu"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "example": "2022-11-14T00:06:03.000000Z"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "example": "2022-11-14T00:06:03.000000Z"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Ao falhar na autorização."
                    }
                }
            },
            "post": {
                "tags": [
                    "Menu"
                ],
                "summary": "Criar menu.",
                "description": "Cria um novo menu.",
                "operationId": "9af9442e517cf9cede4ad30b3810d118",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Novo Menu"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Menu criado com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "name": {
                                            "type": "string",
                                            "example": "Novo Menu"
                                        },
                                        "permalink": {
                                            "type": "string",
                                            "example": "novo-menu"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2022-11-14T00:06:03.000000Z"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2022-11-14T00:06:03.000000Z"
                                        },
                                        "id": {
                                            "type": "integer",
                                            "example": "1"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Ao falhar na autorização."
                    },
                    "409": {
                        "description": "Caso já exista um menu com o mesmo nome."
                    },
                    "422": {
                        "description": "Ao falhar em uma validação.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "name": {
                                                "type": "array",
                                                "items": {}
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/menu/:permalink": {
            "put": {
                "tags": [
                    "Menu"
                ],
                "summary": "Editar usuário.",
                "description": "Edita um usuário.",
                "operationId": "c72be7b1ffa57a8d12a9e04519aa35e0",
                "parameters": [
                    {
                        "name": "permalink",
                        "in": "path",
                        "description": "Permalink do menu",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Novo Menu"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Menu editado com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "name": {
                                            "type": "string",
                                            "example": "Novo Menu"
                                        },
                                        "permalink": {
                                            "type": "string",
                                            "example": "novo-menu"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2022-11-14T00:06:03.000000Z"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2022-11-14T00:06:03.000000Z"
                                        },
                                        "id": {
                                            "type": "integer",
                                            "example": "1"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Ao falhar na autorização."
                    },
                    "409": {
                        "description": "Caso já exista um menu com o mesmo nome."
                    },
                    "422": {
                        "description": "Ao falhar em uma validação.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "name": {
                                                "type": "array",
                                                "items": {}
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Menu"
                ],
                "summary": "Remover menu.",
                "description": "Remove o menu.",
                "operationId": "8d85ffde7bb15b027d87736dc1860864",
                "parameters": [
                    {
                        "name": "permalink",
                        "in": "path",
                        "description": "Permalink do menu",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Menu removido com sucesso."
                    },
                    "401": {
                        "description": "Ao falhar na autorização."
                    },
                    "404": {
                        "description": "Caso o menu não exista."
                    }
                }
            }
        },
...
    }
}